Stamp entire manifest before resolving
Stamp the entire template file before passing it to the resolver. This
allows stamp variables to be used in substitutions, or anywhere
throughout any template.

This is one approach to enabling #351. It has a potential risk in that
anything in a template that matches a stamp variable will be stamped.